requirements.txt
azure-identity
azure-storage-file-share
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-milvus
pymilvus
pypdf
docx2txt
